Bluetooth 배터리 상태 값 로드 코드 리팩토링
